chore: organised dtos into group to make more easier to find
