--- elevate-api/src/main/java/com/elevate/elevateapi/dto/FriendResponse.java
@@ -1,6 +1,0 @@
-package com.elevate.elevateapi.dto;
-
-public record FriendResponse(
-        Long id,
-        String username
-){}
