+++ elevate-api/src/main/java/com/elevate/elevateapi/dto/friends/FriendResponse.java
@@ -1,0 +1,10 @@
+package com.elevate.elevateapi.dto.friends;
+
+import java.time.LocalDate;
+
+public record FriendResponse(
+        Long id,
+        String username
+//        LocalDate date
+//        List<String LiftType, Integer weight, Integer reps> recentLifts
+){}
