+++ elevate-api/src/main/java/com/elevate/elevateapi/dto/friends/IncomingFriendRequests.java
@@ -1,4 +1,4 @@
-package com.elevate.elevateapi.dto;
+package com.elevate.elevateapi.dto.friends;
 
 public record IncomingFriendRequests (
     Long requestId,      // needed to accept / decline
