+++ elevate-api/src/main/java/com/elevate/elevateapi/dto/friends/UserSearchResult.java
@@ -1,4 +1,4 @@
-package com.elevate.elevateapi.dto;
+package com.elevate.elevateapi.dto.friends;
 
 public record UserSearchResult (
         Long id,
