+++ elevate-api/src/main/java/com/elevate/elevateapi/dto/logins/GoogleToken.java
@@ -1,4 +1,4 @@
-package com.elevate.elevateapi.dto;
+package com.elevate.elevateapi.dto.logins;
 
 public record GoogleToken(String idToken) {
 }
