+++ elevate-api/src/main/java/com/elevate/elevateapi/dto/logins/RegisterUserRequest.java
@@ -1,4 +1,4 @@
-package com.elevate.elevateapi.dto;
+package com.elevate.elevateapi.dto.logins;
 
 public record RegisterUserRequest(String username, String password, String email) {
