+++ elevate-api/src/main/java/com/elevate/elevateapi/dto/progresslogs/ProgressLogResponse.java
@@ -1,4 +1,4 @@
-package com.elevate.elevateapi.dto;
+package com.elevate.elevateapi.dto.progresslogs;
 
 import java.time.LocalDate;
