+++ elevate-api/src/main/java/com/elevate/elevateapi/repository/FriendRepository.java
@@ -1,6 +1,6 @@
 package com.elevate.elevateapi.repository;
 
-import com.elevate.elevateapi.dto.FriendResponse;
+import com.elevate.elevateapi.dto.friends.FriendResponse;
 import com.elevate.elevateapi.entity.Friend;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -18,7 +18,7 @@ public interface FriendRepository extends JpaRepository<Friend, Long> {
 
 
     @Query("""
-        select new com.elevate.elevateapi.dto.FriendResponse(
+        select new com.elevate.elevateapi.dto.friends.FriendResponse(
             u.id, u.username)
         from Friend f
         join User u on
