+++ elevate-client/src/app/pages/friends/friends.component.html
@@ -52,7 +52,7 @@
       <h4>{{ f.username }}</h4>
 
       <!-- room in the middle for future info -->
-      <div class="placeholder"> <!-- empty for now --> </div>
+      <div class="placeholder"> Most Recent Lift<!-- empty for now --> </div>
 
       <button class="compare-btn btn"
               (click)="compareWith(f.id)">Compare →</button>
